Added Formula parser from string
Added the FormulaBuilder class which takes in a string and builds up the
Formula Instance.
It goes through each character in the string and sees if it's a number
or starts with '(', if so then it's a molecule.
Then as long as it doesn't encounter a ')', it looks for a capital
letter. If it found one it looks for the other lower case letters to
complete the name and sees if there is a number after it, if so it adds
that also.
If it encounters a '->' then it switches to the right term.