Started working on the solver
I started working on the solver.
The way it works is by creating equations for all the terms. These
equations can be solved to get the coefficients.
The creation of the equations is simple;
Add all the terms on the other side together and substract the terms on
the current side (ofcourse without the term we are currently working
on).

I also added some comments to the logic.